Global replace of single-precision floats with CGFloat casts. (#5718)
Replacing all non-integral, single-precision floating point literals with double-precision values explicitly cast to CGFloat.  For example, `0.1f` should become `(CGFloat)0.1`.

Regex used:
```
([0-9]*\.)([0-9]*?)([1-9]+)([0-9]*?)[fF]
```

PiperOrigin-RevId: 220683126

+++ components/ActionSheet/src/private/MDCActionSheetHeaderView.m
@@ -17,8 +17,8 @@
 #import "MaterialMath.h"
 #import "MaterialTypography.h"
 
-static const CGFloat kTitleLabelAlpha = 0.87f;
-static const CGFloat kMessageLabelAlpha = 0.6f;
+static const CGFloat kTitleLabelAlpha = (CGFloat)0.87;
+static const CGFloat kMessageLabelAlpha = (CGFloat)0.6;
 static const CGFloat kMessageOnlyPadding = 23;
 static const CGFloat kLeadingPadding = 16;
 static const CGFloat kTopStandardPadding = 16;
