+++ components/BottomAppBar/tests/unit/BottomAppBarTests.m
@@ -94,7 +94,7 @@
   CGPoint floatingButtonPosition = self.bottomAppBar.floatingButton.center;
   CGPoint navigationBarPosition = self.bottomAppBar.navBar.frame.origin;
   XCTAssertEqualWithAccuracy(floatingButtonPosition.y + veriticalOffset, navigationBarPosition.y,
-                             0.001f);
+                             (CGFloat)0.001);
 }
 
 #pragma mark - Path test
