+++ components/FeatureHighlight/examples/FeatureHighlightCustomFontsExample.m
@@ -43,7 +43,7 @@
   [self.view addSubview:self.infoLabel];
 
   self.button = [[MDCRaisedButton alloc] init];
-  [self.button setBackgroundColor:[UIColor colorWithWhite:0.1f alpha:1]];
+  [self.button setBackgroundColor:[UIColor colorWithWhite:(CGFloat)0.1 alpha:1]];
   [self.button setTitleColor:[UIColor whiteColor] forState:UIControlStateNormal];
   [self.button setTitle:@"Action" forState:UIControlStateNormal];
   [self.button sizeToFit];
